Implement caching layer to searcher
Now, Searcher accepts a mongo_url parameter that will be used to
store Mongoengine docs. If caching is enabled, search and fetch
will use caching.

The behavior is as follows:
0. __init__
If mongo_url is provided, a connection to the given server
is attempted. If the connection is successful, the 'cached'
property is set to true.

1. Search
It is not served from cache. A search result is only stored.

2. Fetch
If the given UID is present in the cache, it returns the object
from the cache. If the UID is not present, and caching is enabled,
the UID is fetched, the result is stored in the cache and a simple
Dict is supplied from the cached item.

If caching is disabled, an OrderedDict is served.